Разбил слайдер на 2 файла. 1 - Управление размером. 2 - Фильтры
