+++ js/slider.js
@@ -1,27 +1,5 @@
-const fotoPreview = document.querySelector('.img-upload__preview img');
-const sizeFotoPreviev = document.querySelector('.scale__control--value');
-const buttonDecrementSize = document.querySelector('.scale__control--smaller');
-const buttomIncrementSize = document.querySelector('.scale__control--bigger');
-const MAX_SIZE_VALUE = 100;
-const MIN_SIZE_VALUE = 25;
-const SIZE_STEP = 25;
-
-const reducesSize = () => {
-  if (sizeFotoPreviev.value.replace(/\D/g, '') > MIN_SIZE_VALUE) {
-    sizeFotoPreviev.value = `${sizeFotoPreviev.value.replace(/\D/g, '') - SIZE_STEP}%`;
-    fotoPreview.style.transform = `scale(${+sizeFotoPreviev.value.replace(/\D/g, '') / 100})`;
-  }
-};
-
-const increasesSize = () => {
-  if (sizeFotoPreviev.value.replace(/\D/g, '') < MAX_SIZE_VALUE) {
-    sizeFotoPreviev.value = `${+sizeFotoPreviev.value.replace(/\D/g, '') + SIZE_STEP}%`;
-    fotoPreview.style.transform = `scale(${+sizeFotoPreviev.value.replace(/\D/g, '') / 100})`;
-  }
-};
-
-buttonDecrementSize.addEventListener('click', (reducesSize));
-buttomIncrementSize.addEventListener('click', (increasesSize));
+import { fotoPreview } from './changing-picture-size';
+import { sizeFotoPreviev } from './changing-picture-size';
 
 const START_EFFECT_VALUE = 100;
 const sliderElement = document.querySelector('.effect-level__slider');
@@ -49,7 +27,6 @@ noUiSlider.create(sliderElement, {
   step: 1,
   connect: 'lower',
 });
-
 
 effectList.addEventListener('change', (evt) => {
   if (evt.target.checked) {
@@ -153,7 +130,6 @@ effectList.addEventListener('change', (evt) => {
       fotoPreview.style.filter = `brightness(${effectLevelValue.value})`;
     });
   }
-
 });
